requirements.txt
argon2-cffi==25.1.0
cryptography==46.0.4
PyJWT==2.11.0
